REporte cierre de mes funcionando editar, eliminar, nuevo y ultimos detalles
