+++ module/AERSA/view/application/administracion/reportes/cierresinventarios/index.phtml
@@ -1,79 +1,79 @@
 <link href="/application/global/plugins/datatables/datatables.min.css" rel="stylesheet" type="text/css" />
 <link href="/application/global/plugins/datatables/plugins/bootstrap/datatables.bootstrap.css" rel="stylesheet" type="text/css" />
 <div id="my_container">
-    <h3 class="page-title">Cierre inventario</h3>
+    <h3 class="page-title">Cierres inventarios</h3>
     <div class="page-bar">
         <ul class="page-breadcrumb">
             <li>
-                <a href="/administracion/reportes/cierresinventarios">Cierre inventario</a>
+                <a href="/administracion/reportes/cierresinventarios">Cierres inventarios</a>
             </li>
         </ul>
     </div>
-    <?php foreach ($messages->getSuccessMessages() as $message):?>
-        <div class="alert alert-success alert-dismissable">
-            <button class="close" aria-hidden="true" data-dismiss="alert" type="button"></button>
-            <strong>Éxito!</strong>
-            <?php echo $message?>
-        </div>
-    <?php endforeach;?>
-    <div class="portlet box " style="border-left-width: 0px; border-bottom-width: 0px; border-right-width: 0px;">
+    <?php
+    $form->setAttribute('action', '/administracion/reportes/cierresinventarios');
+    echo $this->form()->openTag($form);
+    ?>
+    <div class="portlet box green" style="border-left-width: 0px; border-bottom-width: 0px; border-right-width: 0px;">
         <div class="portlet-body form">
             <div class="form-body">
-                <div class="btn-set pull-right">
-                    <a href="/auditoria/cierresemana/nuevo"><button class="btn default" type="button" style="margin-bottom: 15px;">Nueva cierre inventario</button></a>
+                <div class="row">
+                    <div class="col-md-6">
+                        <?php echo $this->formLabel($form->get('fecha_inicio')) ?>  
+                        <?php echo $this->formText($form->get('fecha_inicio')) ?> 
+                    </div>
+                    <div class="col-md-6">
+                        <?php echo $this->formLabel($form->get('fecha_fin')) ?>  
+                        <?php echo $this->formText($form->get('fecha_fin')) ?> 
+                    </div>
                 </div>
-                <table class="table table-striped table-bordered table-hover order-column" id="datatable"  style="border-bottom-width: 0px;">
-                    <thead>
-                        <tr>
-                            <th class="text"> Fecha </th>
-                            <th class="text"> Alamcen </th>
-                            <th class="text"> Auditor </th>
-                            <th class="text"> Revisado </th>
-                        </tr>
-                    </thead>
-                    <tbody class="text">
-                        <?php $inventario = new Inventariomes();foreach ($inventarios as $inventario) : ?>
-                        <?php 
-                            $revisada = ((bool)$inventario->getInventariomesRevisada()) ? "Revisada":  "No revisada";
-                            $class = ((bool)$inventario->getInventariomesRevisada()) ? "info":  "danger";
-                        ?>
-                        <tr id="<?php echo $inventario->getIdinventariomes()?>">
-                            <td><?php echo $inventario->getInventariomesFecha('d/m/Y') ?></td>
-                            <td><?php echo $inventario->getAlmacen()->getAlmacenNombre() ?></td>
-                            <td><?php echo $inventario->getUsuarioRelatedByIdauditor()->getUsuarioNombre() ?></td>
-                            <td><span class="label label-sm label-<?php echo $class?>"> <?php echo $revisada ?> </span></td>
-                                <td>
-                                    <div class="btn-group">
-                                        <button class="btn btn-xs green dropdown-toggle" type="button" data-toggle="dropdown" aria-expanded="false"> Acciones
-                                            <i class="fa fa-angle-down"></i>
-                                        </button>
-                                        <ul class="dropdown-menu" role="menu">
-                                            <li>
-                                                <a href="/administracion/reportes/cierresinventarios/editar/<?php echo $inventario->getIdinventariomes()?>">
-                                                    <i class="fa fa-pencil"></i> Editar </a>
-                                            </li>
-                                            <li>
-                                                <a href="javascript:;" class="delete_modal">
-                                                    <i class="fa fa-trash"></i> Eliminar </a>
-                                            </li>
-                                        </ul>
-                                    </div>
-                                    
-                                </td>
-                            </tr>                            
-                        <?php endforeach; ?>
-                    </tbody>
-                </table>
+                <br>
+                <div class="row">
+                    <div class="col-md-12">
+                        <p>(*)Campos obligatorios</p>
+                    </div>
+                </div>
+
+            </div>
+            <div class="form-actions">
+                <div class="left col-md-2">
+                    <button disabled class="btn blue" id="generar_pdf" name="generar_pdf" type="submit">
+                        <i class="fa fa-file-pdf-o"></i>
+                        PDF
+                    </button>
+                </div>
+                <div class="left col-md-2">
+                    <button disabled class="btn blue" id="generar_excel" name="generar_excel" type="submit">
+                        <i class="fa fa-file-excel-o"></i>
+                        Excel
+                    </button>
+                </div>
+                <div class="right col-md-8" align="right">
+                    <button disabled class="btn blue" id="mensual_generar" type="button">
+                    <i class="fa fa-check"></i>
+                        Generar
+                    </button>
+                </div>
+            </div>
+            <div class="row">
+                <div id="divtable" class="col-md-12">
+                    <table class="table" id="reporte_table">
+                        <thead>
+                        </thead>
+                        <tbody id="tbody">
+                        </tbody>
+                    </table>
+                </div>
             </div>
         </div>
-        
     </div>
 </div>
 
+<script src="https://code.jquery.com/jquery-1.12.4.js"></script>
+<script src="https://code.jquery.com/ui/1.12.0/jquery-ui.js"></script>
 <script src="/application/global/scripts/datatable.js" type="text/javascript"></script>
 <script src="/application/global/plugins/datatables/datatables.min.js" type="text/javascript"></script>
 <script src="/application/global/plugins/datatables/plugins/bootstrap/datatables.bootstrap.js" type="text/javascript"></script>
-<script src="/application/custom/js/administracion/reportes/cierresinventarios.js" type="text/javascript"></script>
+<script src="/application/custom/js/administracion/reportes/cierresinventarios.js" type="text/javascript"></script>	
 <script>
     var $plugin = $('#my_container').cierresinventarios();
     $plugin.list();
