+++ module/AERSA/view/application/auditoria/cierresinventarios/nuevo.phtml
@@ -1,16 +1,16 @@
 <link href="/application/global/plugins/datatables/datatables.min.css" rel="stylesheet" type="text/css" />
 <link href="/application/global/plugins/datatables/plugins/bootstrap/datatables.bootstrap.css" rel="stylesheet" type="text/css" />
 <div id="my_container">
-    <h3 class="page-title">Inventario cierre de mes</h3>
+    <h3 class="page-title">Inventario cierre de semana</h3>
     <div class="page-bar">
         <ul class="page-breadcrumb">
             <li>
-                <a href="/administracion/reportes/cierresinventarios">Inventario cierre de mes</a>
+                <a href="/auditoria/cierresemana">Inventario cierre de semana</a>
             </li>
         </ul>
     </div>
     <?php
-    $form->setAttribute('action', '/administracion/reportes/cierresinventarios/nuevo');
+    $form->setAttribute('action', '/auditoria/cierresemana/nuevo');
     echo $this->form()->openTag($form);
     ?>
     <div class="portlet box green" style="border-left-width: 0px; border-bottom-width: 0px; border-right-width: 0px;">
@@ -96,7 +96,7 @@
 <script src="/application/global/scripts/datatable.js" type="text/javascript"></script>
 <script src="/application/global/plugins/datatables/datatables.min.js" type="text/javascript"></script>
 <script src="/application/global/plugins/datatables/plugins/bootstrap/datatables.bootstrap.js" type="text/javascript"></script>
-<script src="/application/custom/js/administracion/reportes/cierresinventarios.js" type="text/javascript"></script>
+<script src="/application/custom/js/auditoria/cierresinventarios.js" type="text/javascript"></script>
 <script>
     var $plugin = $('#my_container').cierresinventarios();
     $plugin.new();
